Refactor JSON schema formats
